Implemented a simple model like figure 2.H.
The desired destinations are random positions on the square not on the edge (so people want to actually cross the square). I didn't implement periodic boundary conditions yet either. It would be more realistic, but I think so far it doesn't look that bad. Maybe we should stick with this for now and concentrated on the interacting forces with other pedestrians and obstacles.